third_party/blink/web_tests/external/wpt/service-workers/service-worker/registration-schedule-job.https-expected.txt

This is a testharness.js-based test.
PASS different scriptURL
PASS different updateViaCache
FAIL different type assert_equals: expected (string) "classic" but got (undefined) undefined
Harness: the test ran to completion.
